backend/requirements.txt
Flask==3.0.3
SQLAlchemy==2.0.30
minio==7.1.0
Flask-SQLAlchemy==3.0.3
Flask-Cors==3.0.10
Flask-HTTPAuth==4.8.0
Pillow==9.3.0
psycopg2-binary==2.9.1